move CGeometryManipulator to CGeometryManipulator.h
