+++ include/nbl/asset/utils/CPolygonGeometryManipulator.h
@@ -8,54 +8,12 @@
 #include "nbl/core/declarations.h"
 
 #include "nbl/asset/ICPUPolygonGeometry.h"
+#include "nbl/asset/utils/CGeometryManipulator.h"
 #include "nbl/asset/utils/CQuantNormalCache.h"
 #include "nbl/asset/utils/CQuantQuaternionCache.h"
 
 namespace nbl::asset
 {
-
-// TODO: move to its own header!
-class NBL_API2 CGeometryManipulator
-{
-	public:
-		static inline void recomputeContentHash(const IGeometry<ICPUBuffer>::SDataView& view)
-		{
-			if (!view)
-				return;
-			view.src.buffer->setContentHash(view.src.buffer->computeContentHash());
-		}
-
-		static inline IGeometryBase::SAABBStorage computeRange(const IGeometry<ICPUBuffer>::SDataView& view)
-		{
-			if (!view || !view.composed.isFormatted())
-				return {};
-			auto it = reinterpret_cast<char*>(view.src.buffer->getPointer())+view.src.offset;
-			const auto end = it+view.src.actualSize();
-			auto addToAABB = [&](auto& aabb)->void
-			{
-				using aabb_t = std::remove_reference_t<decltype(aabb)>;
-				for (auto i=0; i!=view.getElementCount(); i++)
-				{
-					typename aabb_t::point_t pt;
-					view.decodeElement(i,pt);
-					aabb.addPoint(pt);
-				}
-			};
-			IGeometryBase::SDataViewBase tmp = {};
-			tmp.resetRange(view.composed.rangeFormat);
-			tmp.visitAABB(addToAABB);
-			return tmp.encodedDataRange;
-		}
-
-		static inline void recomputeRange(IGeometry<ICPUBuffer>::SDataView& view, const bool deduceRangeFormat=true)
-		{
-			if (!view || !view.composed.isFormatted())
-				return;
-			if (deduceRangeFormat)
-				view.composed.rangeFormat = IGeometryBase::getMatchingAABBFormat(view.composed.format);
-			view.composed.encodedDataRange = computeRange(view);
-		}
-};
 
 //! An interface for easy manipulation of polygon geometries.
 class NBL_API2 CPolygonGeometryManipulator
@@ -777,8 +735,6 @@ class CMeshManipulator : public IMeshManipulator
 		CQuantQuaternionCache quantQuaternionCache;
 };
 #endif
-
-// TODO: Utility in another header for GeometryCollection to compute AABBs, deal with skins (joints), etc.
 
 } // end namespace nbl::asset
 #endif
